Refactoring yenier_atkinson_2015
